2425/project/p2/Group_7/Project-SentimentAnalysis/batch-realtime_processing/requirements.txt
praw==7.7.1
kafka-python==2.0.2
python-dotenv==1.0.0
pyspark==3.3.0
findspark==2.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
